requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
jinja2==3.1.2
protobuf==4.25.1
requests==2.31.0
redis==5.0.1 